+++ derive/src/merged_object.rs
@@ -91,6 +91,7 @@ pub fn generate(object_args: &args::MergedObject) -> GeneratorResult<TokenStream
                         keys: ::std::option::Option::None,
                         visible: #visible,
                         is_subscription: false,
+                        rust_typename: ::std::any::type_name::<Self>(),
                     }
                 })
             }