+++ src/types/connection/edge.rs
@@ -108,6 +108,7 @@ where
                 keys: None,
                 visible: None,
                 is_subscription: false,
+                rust_typename: std::any::type_name::<Self>(),
             }
         })
     }