+++ src/types/empty_mutation.rs
@@ -46,6 +46,7 @@ impl Type for EmptyMutation {
             keys: None,
             visible: None,
             is_subscription: false,
+            rust_typename: std::any::type_name::<Self>(),
         })
     }
 }